adjust logic to just reduce the var
