+++ policyengine_us/parameters/gov/contrib/trump/tip_income_tax_exempt/payroll_tax_exempt.yaml
@@ -1,0 +1,8 @@
+description: President Donald Trumps propsal to exempt tip income applies to payroll tax in addition to income tax if this is true.
+metadata:
+  unit: bool
+  period: year
+  label: Trump tip income payroll tax exempt 
+
+values:
+  0000-01-01: false
