+++ policyengine_us/tests/policy/contrib/trump/tip_income_tax_exempt.yaml
@@ -4,31 +4,30 @@
   input:
     gov.contrib.trump.tip_income_tax_exempt.in_effect: true
     tip_income: 20_000
-    adjusted_gross_income: 30_000
-    exemptions: 0
-    taxable_income_deductions: 0
+    irs_gross_income: 30_000
   output:
     taxable_income: 10_000
 
-- name: Reform not in effect
-  period: 2024
-  input:
-    gov.contrib.trump.tip_income_tax_exempt.in_effect: false
-    tip_income: 20_000
-    illicit_income: 30_000
-    exemptions: 0
-    taxable_income_deductions: 0
-  output:
-    taxable_income: 50_000
-
-- name: Reform in effect
+- name: Not Exempt from payroll taxes
   period: 2024
   reforms: policyengine_us.reforms.trump.tip_income_tax_exempt.tip_income_tax_exempt
   input:
     gov.contrib.trump.tip_income_tax_exempt.in_effect: true
+    gov.contrib.trump.tip_income_tax_exempt.payroll_tax_exempt: false
     tip_income: 20_000
-    illicit_income: 30_000
-    exemptions: 0
-    taxable_income_deductions: 0
+    irs_employment_income: 30_000
   output:
-    taxable_income: 30_000
+    taxable_income: 10_000
+    payroll_tax_gross_wages: 30_000
+
+- name: Exempt from payroll taxes
+  period: 2024
+  reforms: policyengine_us.reforms.trump.tip_income_tax_exempt.tip_income_tax_exempt
+  input:
+    gov.contrib.trump.tip_income_tax_exempt.in_effect: true
+    gov.contrib.trump.tip_income_tax_exempt.payroll_tax_exempt: true
+    tip_income: 20_000
+    irs_employment_income: 30_000
+  output:
+    taxable_income: 10_000
+    payroll_tax_gross_wages: 10_000
